Added tabbed view, improved song list tap action
Added a tabbed view to see library as songs or artists or albums, etc. For now only
songs list is viewable, others not implemented. Tapping a song in the library now
creates a play queue consisting of all songs in the selected song's album, ordered by
track no. and the initially playing song set to the one tapped. Makes a lot more sense
in this form.

-MainActivity now loads a LibraryFragment which in turn will create a SongsFragment
  in one of its tabs
-Database service runs only once webpage loads so that the funny gif is guaranteed
  to be seen before library is done initializing :P
-Added helper methid to LibraryHelper for retrieving all songs in an album
-Empty Albums fragment added
-LibraryFragment created that holds a tab host for songs/albums/artists lists
-PlayQueue's new constructor can create a queue from a cursor
-playSong now just gets index from playQueue (currentPlaying) to fix some
  complications
-Fixed next and previous buttons weird placement in PlayerActivity