Album art added, local database removed
Added album art capability. The album art location is stored in a new temporary
cache database for displaying in tracks or in the playerActivity.
Local library database removed completely as it's not really that helpful. Instead
the mediastore will be queried for anything required. One main query runs at
startup to save static cursors for songslist, albumlist, etc.

-Added a playback error message to display on PlayerActivity
-PlayerActivity now shows album art
-Got rid of local library database. All song info is pulled directly from MediaStore.
-Local database stored in cache has been added to hold album art location so that
  mediastore doesnt have to be queried for getting each song's album art location
-SongAdapter changed to query mediastore