Added methods to send DB over I/O stream
Added two methods to Library class that can be used for sending the
library over a connection between devices. Each table in the library is
taken one by one and each of its rows' values along with column names
are put into a JSON object which is then sent over the provided
OutputStream. At the receiving end, the same JSON object is retrieved
from the InputStream and decoded back into a row and inserted into the
same table in the database. Still need to implement sending album art.
I've tried to write the methods as generalized as possible so that
changing columns or adding new tables in the future won't require
updating these methods.

Changes:
Table:
  -Added a constructor to get table name. This can be used by Library
  when accessing a table' name through abstract class 'Table' reference
  -added method to get a cursor of all columns and rows
  -added method to insert values into a new row in table

SongTable, AlbumTable, ArtistTable:
  -calls Table's new constructor
  -uses insertValues()

Library:
  -Forgot to set querySongs() ordering back to song title. (was set to
  album name for testing in previous commit)
  -Added two methods for sending/receiving the entire library database
  over input/output streams as JSON objects