Methods updated to send album art as well after sending library DB
The methods for sending and receiving the music library have been
updated to send album art after the sending the database.
Also some statements have been added so that the sender pauses and lets
the receiver process data in between. I'm not so sure if the sender is
actually overloading the receiver, but there were some bugs where the
receiver would only receive a fraction of the album art, with mixing
data of 2 images into one file. Adding these "wait" and "ready"
statements seemed to solve the issue though.

Library:
  -Updated sendOverStream() and getFromStream() to send/receive album
  art as well.
  -Added response/ready statements to let receiver process data