Limited testing version for testing library sharing
Splash activity was modified for testing the library sharing methods.
When the app is opened, the user can choose to open the app in normal mode
as before to play local songs, or choose the app to act as a sender or
receiver. The sender just has to wait. The receiver takes the IP address
of the sender through the text field (user has to enter the ip of the
sender device). The receiver then connects to the sender through
sockets and the sender's library starts transferring to it using the
send/get methods in Library class.
Once transfer completes, the sender sees their app as usual but the
receiver sees the senders song list and album list instead of their
local one.

SplashActivity:
  -Added buttons to choose one of 3 modes: Sender, Receiver or normal.
  Normal will open the app as it was before.

LibraryService:
  -Library database is stored in internal SD card instead of cache
  allowing opening by other apps. Did that for debugging/testing
  -Second database "testlib" also created in internal SD. If app is
  opened in receiver mode, the sender device's music library is saved
  here.
  -IP address taken from SplashActivity in Receiver mode is used to
  connect with sender on the network.
  -If in receiver mode, the music and album cursor points to testlib
  instead of the local library database.

AndroidManifest.xml:
  -Added internet permission for using sockets.