App aware of new members joining. Library exchanging implemented at joining
App is now aware of other instances of this app running on the Wifi P2P
network.
The architecture I chose has NetworkService at the bottom, using
Connection for all socket operations. Above NetworkService is ShareGroup
that interfaces with the fragments.
The NetworkService takes incoming requests, breaks them down into an
"action" and possible input data and passes that to ShareGroup.
ShareGroup sends a response data, if needed, and the NetworkService
sends that back to the request initiator.
Tested with two devices. They can connect to each other and exchange
libraries successfully. Very limited though as it only works if
following strict steps. Special conditions and errors haven't yet been
considered or fixed. Also disconnection from a group isn't implemented
yet.

I'm happy with the overall architecture I developed to help implement
all the new code. However, it looks pretty messy and doesn't handle
exceptional use cases well. Hope to clean things up in the future.

Changes:
SplashActivity:
  -Initializes library variables for local and remote library database
  location and remote library's album art cache location

Connection:
  -A listener connection (server socket) is an inner class now, and
  instantiated separately. From this class, new Connection objects can
  be created for incoming requests.
  -Has methods for sending and receiving data. For now, can transfer
  Strings, integers and (default case) just a serializable object.
  -Contains a wait/ready system for cases where one device might only
  send multiple sets of data without receiving. This method prevents
  any lost objects and forces the writer to pause and wait for the
  receiver to be ready for the next data.

NetworkService:
  -NetworkManager renamed and made into a service.
  -Totally revamped to handle new architecture for establishing
  connection between devices. Honestly, I've lost track of the shear
  amount of changes and so has git apparently.

ShareGroup:
  -Heavily modified to to logically handle new GlidePlayer members
  joining the group

ConnectivityFragment:
  -Uses a listener to wait until the ShareGroup and NetworkService are
  ready before sending a ShareGroup object to the respective fragment
  (CreateGroup/JoinGroup). This had to be done because of the
  asynchronous nature of NetworkService

PlayerFragment:
  -Changed to talk with the PlayerService through a binder. This
  eliminates the need for a Handler.

AlbumTable,SongTable,ArtistTable:
  -Moved initialize to the abstract class Table.

PlayerService:
  -Changed to use a binder for interacting with any other object.
  The binder takes requests and performs it asynchronously using the
  service methods.

New classes:
WifiP2pBroadcastReceiver:
  -Used for handling WifiP2p API's broadcasts on behalf of the
  NetworkService

EasyHandler:
   -A class for easily running tasks asynchronously on a separate
   thread, or for running tasks on the main thread from within a
   separate thread
   -It can create HandlerThreads when needed and keeps them running so
   that more tasks can be posted onto it by assigning it to a thread
   name

New Listeners:
GroupConnectionListener: For connection sucess, failure or to inform the
fragment that wifi connection is complete but library still need to be
exchanged

GroupMemberListener: For notifying when a member joins or leaves the
group. For now only joining is functional

RequestListener: Called when a new request reaches for the ShareGroup
class to handle. The listener takes an object as response data from the
listener as a return value (can be null if no response data required)

ResponseListener: Used for when the ShareGroup sends a new request.
Methods in this interface are called either when the request succeeds
(and response data is provided) or when the request fails