3 and more devices can connect and share libraries.
Turns out the reason 3+ devices couldn't connect was because the
clientMap hashmap could not be sent over Object Stream due to
serialization issues. Now, the client map is broken down into a JSON
object and reconstructed into a new client map at the receiving end.
Essentially, any number of devices can now connect and see and play
each other's music library. I've only been able to test it with 3 and it
seems to work. Will test with 4 or more later and see if there are any bugs.

Changes:
NetworkService:
  -Client class can be constructed from and deconstructed to a JSON
  object.
  -Methods sendClientMap() and getClientMap() can be used to send
  the client map to a new device requesting establishment. They change
  the client map to JSON objects and send the JSON object as string over
  the socket.