Continuous queue playback and track seeking added
When a song finishes playing, the next song starts. When the last song
in the queue finishes playing, playback just stops. And finally, track
seeking has been implemented. Some changes still need to be made
and I'm considering a different logic in some cases, but I'll push this
commit anyway, in case I dont get back to the possible changes in a
while. This is good enough for now

Changes:
PlayerFragment:
  -Tested and re-implemented service binding style for player service.
  -Seek bar implemented
  -Implemented new callbacks (see player service changes)

PlayerService:
  -Service binding tested and implemented.
  -When a song completes, next song is started
  -When last song in queue is reached, playback stops
  -New callbacks added to listener to reflect these changes.
  onTrackAutoChanged notifies listening classes that the song changed
  (next track), onSongStopped notifies when a song stops playing (only
  called when last song in queue finishes), onSeekUpdate is called
  regularly when the song seeks, so that listening classes can update a
  seekbar, if they have one.
  -Multiple song listeners can be registered.

MusicPlayer:
  -OnCompletionListener used
  -A new thread constantly runs to check and notify of song seek, but
  the thread only runs when actually needed.

PlayQueue:
  -Added a method to check if current song is the last song.