Renamed package and version, Clear remote table entries
Renamed the package to com.teefourteen.glideplayer.
Changed version name to 0.1
Removed creation of Remote libraries directory since it's stored in the
local database.
When a user disconnects from the group, all remote table entries from
that user is deleted from the database. If disconnected from the group,
all entries in the remote table are deleted.

Changes:
Library:
  -deleteUser() removes all entries of that user from the remote tables.
  -clearRemoteTables() deletes all rows from all remote tables.

ShareGroup:
  -Calls new methods in Library.