Small fixes
Player notification is made dismissable anytime the song is paused, even if still inside the app whereas previously, it could only be dismissed if the app is closed (ie, PlayerService is unbound).
Left margin increased for library spinner's content
Added a stop and release call for the music player on closing the object.

Changes:
MusicPlayer:
  -Stops and releases music player on close() if it is prepared.

PlayerService:
  -Notification's ongoing flag is removed anytime the song is paused

library_spinner.xml:
  -Left margin for the username increased.